+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @under-control/forms
 
+## 1.6.0
+
+### Minor Changes
+
+- Add related inputs helper
+
+### Patch Changes
+
+- Updated dependencies
+  - @under-control/inputs@1.6.0
+  - @under-control/core@1.6.0
+  - @under-control/validate@1.6.0
+
 ## 1.5.2
 
 ### Patch Changes